Extract keyboard handling to separate file
